+++ packages/json_rpc_codegen_builder/lib/src/generators/client/stream_builder_mixin.dart
@@ -206,7 +206,8 @@ base mixin ClientStreamBuilderMixin
         .asA(Types.$StreamController(streamType.toReference()).asNullable(true))
         .nullSafeProperty('add')
         .call([
-          fromJson(
+          methodFromJson(
+            method,
             streamType,
             streamType.isNullableType
                 ? params