+++ packages/json_rpc_codegen_builder/lib/src/generators/common/method_mapper_mixin.dart
@@ -3,12 +3,14 @@ import 'package:analyzer/dart/element/type.dart';
 import 'package:code_builder/code_builder.dart';
 import 'package:collection/collection.dart';
 import 'package:dart_test_tools/code_gen.dart';
-import 'package:json_rpc_codegen/json_rpc_codegen.dart' show ParamName;
+import 'package:json_rpc_codegen/json_rpc_codegen.dart'
+    show RpcMethod, RpcParam;
 import 'package:meta/meta.dart';
 import 'package:source_gen/source_gen.dart';
 
 import '../../extensions/parameter_extensions.dart';
-import '../../readers/name_reader.dart';
+import '../../readers/rpc_method_reader.dart';
+import '../../readers/rpc_param_reader.dart';
 import 'annotations.dart';
 
 @internal
@@ -31,17 +33,18 @@ enum ReturnKind { notification, request, stream }
 base mixin MethodMapperMixin {
   @protected
   String rpcMethodName(MethodElement method, [String invocationSuffix = '']) =>
-      '${NameReader.methodName(method)?.name ?? method.name!}$invocationSuffix';
+      '${RpcMethodReader.read(method)?.name ?? method.name!}$invocationSuffix';
 
   @protected
   String rpcParamName(FormalParameterElement param) =>
-      NameReader.paramName(param)?.name ?? param.name!;
+      RpcParamReader.read(param)?.name ?? param.name!;
 
   @protected
   ({DartType type, ReturnKind kind}) getReturnType(MethodElement method) {
     final returnType = method.returnType;
 
     if (returnType is VoidType) {
+      _validateNoConverters(method);
       return (type: returnType, kind: .notification);
     }
 
@@ -64,6 +67,10 @@ base mixin MethodMapperMixin {
     }
 
     final futureType = (returnType as InterfaceType).typeArguments.single;
+
+    if (futureType is VoidType) {
+      _validateNoConverters(method);
+    }
 
     if (futureType.isDartCoreType ||
         futureType.isDartCoreSymbol ||
@@ -97,12 +104,12 @@ base mixin MethodMapperMixin {
     }
 
     for (final param in method.formalParameters.where((e) => e.isPositional)) {
-      if (NameReader.paramName(param) != null) {
+      if (RpcParamReader.read(param)?.name != null) {
         throw InvalidGenerationSourceError(
-          'The $ParamName annotation can only be used on named parameters, as '
-          'positional parameters are transmitted as a list.',
+          'The name of the $RpcParam annotation can only be used on named '
+          'parameters, as positional parameters are transmitted as a list.',
           element: param,
-          todo: 'Remove the annotation or make the parameter named.',
+          todo: 'Remove the name or make the parameter named.',
         );
       }
     }
@@ -145,6 +152,18 @@ base mixin MethodMapperMixin {
       );
     buildMethod(b);
   });
+
+  /// Ensures [method] has no conversion functions, as it returns nothing.
+  void _validateNoConverters(MethodElement method) {
+    if (RpcMethodReader.read(method)?.hasConverters ?? false) {
+      throw InvalidGenerationSourceError(
+        'The conversion functions of the $RpcMethod annotation cannot be used '
+        'on methods that do not return a value.',
+        element: method,
+        todo: 'Remove the fromJson and toJson functions.',
+      );
+    }
+  }
 
   Never _throwNotAFuture(MethodElement method, DartType displayType) =>
       throw InvalidGenerationSourceError(